pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "urbanflow"
version = "0.1.0"
description = "UrbanFlow MVP transit network optimization pipeline"
readme = "README.md"
authors = [
  { name = "UrbanFlow Team" }
]
license = { text = "Proprietary" }
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "geopandas>=0.14",
  "numpy>=1.26",
  "scipy>=1.11",
  "networkx>=3.2",
  "folium>=0.15",
  "click>=8.1",
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29",
  "gtfs-kit>=5.2",
  "partridge>=1.2",
  "pyyaml>=6.0",
  "typer>=0.12"
]

[project.optional-dependencies]
api = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.29"
]

[project.scripts]
urbanflow = "urbanflow.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra"
pythonpath = ["src"]
